Added segmentGPU to _SegNet

+++ src/Utility/util.h
@@ -29,7 +29,8 @@ inline uint64_t get_time_usec()
 {
 	static struct timeval _time_stamp;
 	gettimeofday(&_time_stamp, NULL);
-	return _time_stamp.tv_sec * 1000000 + _time_stamp.tv_usec;
+//	printf("%ld\n",(uint64_t)_time_stamp.tv_sec * (uint64_t)1000000 + _time_stamp.tv_usec);
+	return (uint64_t)_time_stamp.tv_sec * (uint64_t)1000000 + _time_stamp.tv_usec;
 }
 
 /*
